pyproject.toml
[project]
name = "ai-workflow-engine"
version = "0.1.0"
description = "Framework Python d'orchestration d'agents IA pour workflows de développement structurés"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "B68682" }]

dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "transitions>=0.9",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
]

[project.scripts]
ai-workflow = "ai_workflow.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engine/src/ai_workflow"]

[tool.ruff]
src = ["engine/src"]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["engine/tests"]
pythonpath = ["engine/src"]

[tool.coverage.run]
source = ["ai_workflow"]
